CD-161: Cluster nodes run as backup & prepared mission control for failover
- Added support for MissionParameters (parameters with values that can be changed at runtime) to MissionControl
- Updated DebugCanvas to display Role / ID / Render ID
- Centralized NodeRole into a new ClusterDisplay.Shared assembly so that both ClusterDisplay and Capcom can use it.
- Fixed bug in the UI causing it to sometime falsely report unaccounted for changes from the server in the launch configuration.
- L.aunchpad can now set its state to over instead of idle when aborting (to have a more similar behaviorbetween cases where a launchable finishes by itself vs is forcefully terminated).
- Centralized strings identifying collections in incrementalCollectionsUpdate.
- Centralized strings identifying objects in objectsUpdate.
- Moved Shared.LaunchCatalog.Constraints to Shared (since it is not also used for MissionParameters).
- Moved update version numbers out of MissionControlMirror to another private to Application object.

+++ TestProjects/MissionControlTests/Assets/Tests/Capsule/SendMessagesTests.cs
@@ -161,7 +161,7 @@ namespace Unity.ClusterDisplay.MissionControl.Capsule
 
             // Validate the received message
             Assert.That(capsuleStatusMessageId, Is.EqualTo(MessagesId.CapsuleStatus));
-            Assert.That(capsuleStatusMessage.NodeRole, Is.EqualTo((byte)NodeRole.Repeater));
+            Assert.That(capsuleStatusMessage.NodeRole, Is.EqualTo(NodeRole.Repeater));
             Assert.That(capsuleStatusMessage.NodeId, Is.EqualTo(28));
             Assert.That(capsuleStatusMessage.RenderNodeId, Is.EqualTo(42));
